+++ CHANGELOG.md
@@ -25,3 +25,5 @@
 - Add `ResultEqualityComparer<T>`.
 
 - Add additional overloads for `Result<T>.Equals` which allow specifying an equality comparer to use for comparing values.
+
+- Rename `GetValueOrDefault(T @default)` and `GetValueOrDefault(Func<T> getDefault)` to `GetValueOr`.